refactor: move cli files
