+++ packages/vitest/src/node/cache/index.ts
@@ -4,7 +4,7 @@ import { findUp } from 'find-up'
 import { resolve } from 'pathe'
 import { loadConfigFromFile } from 'vite'
 import { configFiles } from '../../constants'
-import type { CliOptions } from '../cli-api'
+import type { CliOptions } from '../cli/cli-api'
 import { slash } from '../../utils'
 import { FilesStatsCache } from './files'
 import { ResultsCache } from './results'
