+++ packages/vitest/src/node/cli/cli-config.ts
@@ -1,7 +1,7 @@
 import { defaultBrowserPort, defaultPort } from '../../constants'
 import type { ApiConfig } from '../../types/config'
 import type { ForksOptions, ThreadsOptions, VmOptions, WorkerContextOptions } from '../../types/pool-options'
-import type { CliOptions } from '../cli-api'
+import type { CliOptions } from './cli-api'
 
 type NestedOption<T, V = Extract<T, Record<string, any>>> = V extends never | RegExp | unknown[]
   ? never
