+++ packages/vitest/src/node/index.ts
@@ -2,7 +2,7 @@ export type { Vitest } from './core'
 export type { WorkspaceProject } from './workspace'
 export { createVitest } from './create'
 export { VitestPlugin } from './plugins'
-export { startVitest } from './cli-api'
+export { startVitest } from './cli/cli-api'
 export { registerConsoleShortcuts } from './stdin'
 export type { GlobalSetupContext } from './globalSetup'
 export type { WorkspaceSpec, ProcessPool } from './pool'
